+++ chatter/chat-app/src/app/chat/chat-list/chat-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { TokenService } from 'src/app/services/token.service';
+import { TokenService } from 'src/app/shared/services/token.service';
 import { UserService } from 'src/app/streams/services/user.service';
 import { PayloadData } from 'src/app/shared/interfaces/jwt-payload.interface';
 import { User, ChatList } from 'src/app/shared/interfaces/user.interface';
