+++ chatter/chat-app/src/app/chat/chat-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { ChatComponent } from './chat.component';
-import { AuthGuard } from '../guards/auth.guard';
+import { AuthGuard } from '../shared/guards/auth.guard';
 import { MessageComponent } from './message/message.component';
 import { ChatListComponent } from './chat-list/chat-list.component';
 import { NotificationsComponent } from '../chat/notifications/notifications.component';
