+++ chatter/chat-app/src/app/shared/interceptors/jwt.interceptor.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpInterceptor, HttpRequest, HttpHandler, HttpEvent } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { TokenService } from '../services/token.service';
-import { CustomRequest } from '../shared/interfaces/custom-request.interface';
+import { CustomRequest } from '../interfaces/custom-request.interface';
 
 @Injectable()
 export class JwtInterceptor implements HttpInterceptor {
