+++ chatter/chat-app/src/app/shared/services/application-state.service.ts
