+++ chatter/chat-app/src/app/shared/side-nav/side-nav.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { TokenService } from '../../services/token.service';
+import { TokenService } from '../services/token.service';
 import { Router } from '@angular/router';
 
 @Component({
