+++ chatter/chat-app/src/app/streams/streams.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { ApplicationStateService } from '../services/application-state.service';
-import { TokenService } from '../services/token.service';
+import { ApplicationStateService } from '../shared/services/application-state.service';
+import { TokenService } from '../shared/services/token.service';
 import { PayloadData } from '../shared/interfaces/jwt-payload.interface';
 
 @Component({
