+++ chatter/chat-app/src/app/streams/users/followers/followers.component.ts
@@ -4,7 +4,7 @@ import { PayloadData } from '../../../shared/interfaces/jwt-payload.interface';
 import { UserFollowed } from '../../interfaces/user-followed.interface';
 import * as _ from 'lodash';
 import { UserFollowing } from '../../interfaces/user-following.interface';
-import { TokenService } from '../../../services/token.service';
+import { TokenService } from '../../../shared/services/token.service';
 import { ActivatedRoute } from '@angular/router';
 import { NzNotificationService } from 'ng-zorro-antd';
 import { User } from '../../../shared/interfaces/user.interface';
