+++ chatter/chat-app/src/app/streams/users/following/following.component.ts
@@ -4,7 +4,7 @@ import { UserFollowed } from '../../interfaces/user-followed.interface';
 import { UserService } from '../../services/user.service';
 import { PayloadData } from '../../../shared/interfaces/jwt-payload.interface';
 import * as _ from 'lodash';
-import { TokenService } from '../../../services/token.service';
+import { TokenService } from '../../../shared/services/token.service';
 import { ActivatedRoute } from '@angular/router';
 import { NzNotificationService, NzConfigService } from 'ng-zorro-antd';
